move and refactor character.c / .h
